+++ InternetShop/Shop/ShopItemExtended.cs
@@ -3,7 +3,7 @@
     public class ShopItemExtended : ShopItem
     {
         public string ItemDescription { get; set; }
-        public ShopItemExtended(string itemName, double itemPrice, string itemDescription) : base(itemName, itemPrice)
+        public ShopItemExtended(string itemName, int itemPrice, string itemDescription) : base(itemName, itemPrice)
         {
             ItemDescription = itemDescription;
         }
